:recycle: Refactor 패키지 구조 변경 및 수정용 dto 추가
