+++ src/main/java/com/nitrogen/domain/expense/converter/ExpenseConverter.java
@@ -1,7 +1,5 @@
 package com.nitrogen.domain.expense.converter;
 
-import com.nitrogen.domain.expense.dto.ExpenseDetailsDTO;
-
 public class ExpenseConverter {
 
 }
