+++ src/main/java/com/nitrogen/domain/expense/dto/category/CategoryUpdateRequestDTO.java
@@ -1,0 +1,14 @@
+package com.nitrogen.domain.expense.dto.category;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+public class CategoryUpdateRequestDTO {
+    private String name;
+}
