+++ src/main/java/com/nitrogen/domain/expense/dto/expense/ExpenseDetailsDTO.java
@@ -1,4 +1,4 @@
-package com.nitrogen.domain.expense.dto;
+package com.nitrogen.domain.expense.dto.expense;
 
 import com.nitrogen.domain.expense.entity.enums.EmotionType;
 import jakarta.validation.constraints.NotBlank;
